refactor,fix: Rewrite MVC model and fix load error
