+++ asset/index.php
@@ -1,0 +1,2 @@
+<?php
+?>
