--- asset/php/controller/messages/fetch.php
@@ -1,11 +1,0 @@
-<?php
-
-require 'MessageController.php';
-
-$message=new MessageController();
-$data = $message->getData();
-
-R::close();
-exit();
-
-?>
