--- asset/php/index.php
@@ -1,5 +1,0 @@
-<?php
-
-include './db/init.php' ;
-
-?>
