--- asset/php/models/helper/auxiliarMmethods.php
@@ -1,13 +1,0 @@
-<?php
-
-trait globalFonc
-{
-    public function sendTime()
-    {
-        date_default_timezone_set("Asia/Tehran");
-        $time = time();
-        return $time;
-    }
-}
-
-?>
