+++ asset/src/controller/messages/insert.php
@@ -9,13 +9,19 @@ if (!empty($messageText)) {
     $message = new MessageController();
     try {
         $message->setData($messageText, 404, $chat_name);
+        header('Content-Type: application/json');
+        http_response_code(200);
+        echo json_encode([
+            'status' => 'success',
+        ]);
     } catch (Exception $err) {
-        echo "hi";
+        header('Content-Type: application/json');
         http_response_code(500);
         echo json_encode([
             'status' => 'failed',
             'message' => $err->getMessage()
         ]);
+        error_log($err->getMessage() . "\n", 3, "err.txt");
     }
 }
 R::close();
