+++ asset/src/view/index.html
@@ -3,24 +3,13 @@
   <head>
     <meta charset="UTF-8" />
     <title>Chat</title>
-    <link rel="stylesheet" href="./asset/style/style_chat.css" />
-    <link rel="stylesheet" href="./asset/style/style_chat_folders.css" />
-    <link rel="stylesheet" href="./asset/style/style_chat_chatlist.css" />
-    <link rel="stylesheet" href="./asset/style/style_chat_dialog.css" />
-    <link rel="stylesheet" href="./asset/style/style_icons_and_fonts.css" />
-    <link rel="stylesheet" href="./asset/style/style_add_contacts.css" />
+    <link rel="stylesheet" href="../../style/style_chat.css" />
+    <link rel="stylesheet" href="../../style/style_chat_folders.css" />
+    <link rel="stylesheet" href="../../style/style_chat_chatlist.css" />
+    <link rel="stylesheet" href="../../style/style_chat_dialog.css" />
+    <link rel="stylesheet" href="../../style/style_icons_and_fonts.css" />
+    <link rel="stylesheet" href="../../style/style_add_contacts.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-    <script
-      src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js"
-      integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
-      crossorigin="anonymous"
-    ></script>
-    <script
-      src="https://code.jquery.com/jquery-3.3.1.min.js"
-      integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="
-      crossorigin="anonymous"
-    ></script>
   </head>
 
   <body onload="refreshChatlist()">
@@ -78,7 +67,7 @@
           <div class="dialog__header">
             <div class="dialog__header-right">
               <img
-                src="./asset/image/user.png"
+                src="../../image/user.png"
                 alt="profile"
                 class="message__img"
               />
@@ -240,6 +229,6 @@
       integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
       crossorigin="anonymous"
     ></script>
-    <script src="./asset/js/pageReaction.js"></script>
+    <script src="../../js/pageReaction.js"></script>
   </body>
 </html>
