+++ jsonFiles/Contacts.json
@@ -6,7 +6,7 @@
         "phone":"",
         "fullNname": "php Channel",
         "userName":"@php_channel",
-        "profile": "./asset/image/user.png",
+        "profile": "../../image/user.png",
         "chatType": "channel",
         "sender": "php Channel",
         "lastMessage": "php...",
@@ -22,7 +22,7 @@
         "phone":"09399652583",
         "fullNname": "Donya Dastkin",
         "userName":"@donyaa-583",
-        "profile": "./asset/image/Donya_Dstkin.jpg",
+        "profile": "../../image/Donya_Dstkin.jpg",
         "chatType": "pv",
         "sender": "Donya Dastkin",
         "lastMessage": "Hiii",
@@ -37,7 +37,7 @@
         "phone":"",
         "fullNname": "farawin",
         "userName":"@farawin",
-        "profile": "./asset/image/farawin_g.jpeg",
+        "profile": "../../image/farawin_g.jpeg",
         "chatType": "group",
         "sender":"Donya",
         "lastMessage": "cherttt",
